SAM/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0
shapely>=2.0.0
geopandas>=0.14.0
rasterio>=1.3.0
torch>=2.0.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
